quickfix_users/backend/requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
Flask-Bcrypt
Flask-Cors
PyJWT
python-dotenv